Move ruby diagnostics functionality to build-utils.
